Fix teh UI HTML view and LWC wrappers, provide raw component and LEX Card as demo wrapper
