+++ force-app/main/default/lwc/escomboChildList/escomboChildList.html
@@ -1,0 +1,20 @@
+<template>
+    <template if:false={showOption}>
+        <lightning-combobox
+            label="Select Reason"
+            value={variant}
+            onchange={variantChange}
+            options={variantOptions}
+        >
+        </lightning-combobox>
+    </template>
+        <template if:true={showOption}>
+            <lightning-input
+            type="text"
+            onchange={handleKeyChange}
+            class="slds-var-m-bottom_small"
+            label="New Value"
+            value={searchKey}
+            ></lightning-input>
+        </template>
+</template>
