+++ force-app/main/default/lwc/escomboMainList/escomboMainList.html
@@ -1,0 +1,10 @@
+<template>
+    <!-- Basic combo component derive data list -->
+    <lightning-combobox
+        label="Select Reason"
+        value={variant}
+        onchange={variantChange}
+        options={variantOptions}
+    >
+    </lightning-combobox>
+</template>
