+++ force-app/main/default/objects/List_Metadata__c/listViews/All.listView-meta.xml
@@ -1,6 +1,13 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <ListView xmlns="http://soap.sforce.com/2006/04/metadata">
     <fullName>All</fullName>
+    <columns>NAME</columns>
+    <columns>Reason_Label__c</columns>
+    <columns>Reason_Value__c</columns>
+    <columns>Child_Label__c</columns>
+    <columns>Child_Value__c</columns>
+    <columns>Active__c</columns>
+    <columns>OBJECT_ID</columns>
     <filterScope>Everything</filterScope>
     <label>All</label>
 </ListView>
