+++ force-app/main/default/tabs/Example_List_Page.tab-meta.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<CustomTab xmlns="http://soap.sforce.com/2006/04/metadata">
+    <description>Created by Lightning App Builder</description>
+    <flexiPage>Example_List_Page</flexiPage>
+    <label>Example List Page</label>
+    <motif>Custom17: Sack</motif>
+</CustomTab>
